+++ conwaymd/employables.py
@@ -1862,6 +1862,7 @@ class SpecifiedLinkReplacement(
 class ReferencedLinkReplacement(
     ReplacementWithAllowedFlags,
     ReplacementWithAttributeSpecifications,
+    ReplacementWithHrefReplacements,
     ReplacementWithProhibitedContent,
     Replacement,
 ):
@@ -1874,6 +1875,7 @@ class ReferencedLinkReplacement(
     - queue_position: (def) NONE | ROOT | BEFORE #«id» | AFTER #«id»
     - allowed_flags: (def) NONE | «letter»=«FLAG_NAME» [...]
     - attribute_specifications: (def) NONE | EMPTY | «string»
+    - href_replacements: (def) NONE | #«id» [...]
     - prohibited_content: (def) NONE | BLOCKS | ANCHORED_BLOCKS
     ````
     """
@@ -1893,6 +1895,7 @@ class ReferencedLinkReplacement(
             'queue_position',
             'allowed_flags',
             'attribute_specifications',
+            'href_replacements',
             'prohibited_content',
         )
 
@@ -1988,7 +1991,10 @@ class ReferencedLinkReplacement(
             else:
                 combined_attribute_specifications = href_title_referenced_attribute_specifications
 
-            attributes_sequence = build_attributes_sequence(combined_attribute_specifications, use_protection=True)
+            attributes_sequence = build_attributes_sequence(combined_attribute_specifications,
+                                                            use_protection=True,
+                                                            href_replacements=self._href_replacements,
+                                                            enabled_flag_names=enabled_flag_names)
 
             substitute = f'<a{attributes_sequence}>{content}</a>'